Add Server_Client Code
This is the code I have so far, it counts the bytes and echo-s but it
doesn’t have continuous stream yet which I’m working on and it don’t
have multiple instances (which is what Matthew needs).  The current
port is 1101 (which it will automatically connect to if you need it
to).  This is just a start, I will keep updating it as I work on it
throughout the day.