Reestablishing my code
I found a better example of how to do the server_client code.  I have a
single connection working now with better documentation (i embedded it
in the code so you can follow it, it’s pretty straight forward).  But I
have to incorporate the p_threading in again… Sco Ducks.